feat(scribe): gate recording, recover failed, refactor scribe
