+++ src/components/patient/scribeVisuals.jsx
@@ -1,0 +1,235 @@
+import { Box } from "@chakra-ui/react";
+import { keyframes } from "@emotion/react";
+
+// Animation keyframes
+const pulse = keyframes`
+  0%, 100% {
+    opacity: 0.3;
+    transform: scale(1);
+  }
+  50% {
+    opacity: 0.6;
+    transform: scale(1.1);
+  }
+`;
+
+const lavaBlob1 = keyframes`
+  0%, 100% { transform: translate(0, 0) scale(1); }
+  33% { transform: translate(6px, -4px) scale(1.2); }
+  66% { transform: translate(-4px, 5px) scale(0.9); }
+`;
+
+const lavaBlob2 = keyframes`
+  0%, 100% { transform: translate(0, 0) scale(1); }
+  33% { transform: translate(-5px, 4px) scale(0.85); }
+  66% { transform: translate(4px, -3px) scale(1.15); }
+`;
+
+const lavaBlob3 = keyframes`
+  0%, 100% { transform: translate(0, 0) scale(1); }
+  50% { transform: translate(3px, 6px) scale(1.1); }
+`;
+
+// Swirling orb animations for loading state
+const swirlOrbit = keyframes`
+  0% { transform: rotate(0deg); }
+  100% { transform: rotate(360deg); }
+`;
+
+const swirlOrbitReverse = keyframes`
+  0% { transform: rotate(360deg); }
+  100% { transform: rotate(0deg); }
+`;
+
+const corePulse = keyframes`
+  0%, 100% { transform: scale(1); opacity: 0.8; }
+  50% { transform: scale(1.15); opacity: 1; }
+`;
+
+const orbPulse = keyframes`
+  0%, 100% { opacity: 0.7; transform: scale(1); }
+  50% { opacity: 1; transform: scale(1.2); }
+`;
+
+const fadeIn = keyframes`
+  0% { opacity: 0; transform: scale(0.8); }
+  100% { opacity: 1; transform: scale(1); }
+`;
+
+// Lava lamp blob layers for idle button
+export const LavaBlobs = () => (
+    <Box
+        position="absolute"
+        top={0}
+        left={0}
+        right={0}
+        bottom={0}
+        borderRadius="full"
+        overflow="hidden"
+        pointerEvents="none"
+    >
+        <Box
+            position="absolute"
+            top={0}
+            left={0}
+            right={0}
+            bottom={0}
+            bg="linear-gradient(135deg, #BF360C, #E65100)"
+        />
+        <Box
+            position="absolute"
+            top="-10%"
+            left="-10%"
+            w="70%"
+            h="70%"
+            borderRadius="50%"
+            bg="radial-gradient(circle, #FF6B35 0%, transparent 70%)"
+            animation={`${lavaBlob1} 4s ease-in-out infinite`}
+            opacity={0.8}
+        />
+        <Box
+            position="absolute"
+            top="40%"
+            left="30%"
+            w="60%"
+            h="60%"
+            borderRadius="50%"
+            bg="radial-gradient(circle, #FF8A50 0%, transparent 70%)"
+            animation={`${lavaBlob2} 3s ease-in-out infinite`}
+            opacity={0.7}
+        />
+        <Box
+            position="absolute"
+            top="20%"
+            left="40%"
+            w="50%"
+            h="50%"
+            borderRadius="50%"
+            bg="radial-gradient(circle, #FFB74D 0%, transparent 20%)"
+            animation={`${lavaBlob3} 5.5s ease-in-out infinite`}
+            opacity={0.5}
+        />
+    </Box>
+);
+
+// Internal pulsing glow for recording state
+export const InternalGlow = () => (
+    <Box
+        position="absolute"
+        top={0}
+        left={0}
+        right={0}
+        bottom={0}
+        borderRadius="full"
+        overflow="hidden"
+        pointerEvents="none"
+    >
+        <Box
+            position="absolute"
+            top={0}
+            left={0}
+            right={0}
+            bottom={0}
+            borderRadius="full"
+            bg="radial-gradient(circle, rgba(229,62,62,0.4) 0%, rgba(229,62,62,0.1) 50%, transparent 70%)"
+            animation={`${pulse} 1.5s ease-in-out infinite`}
+        />
+    </Box>
+);
+
+// Swirling loading orb component
+export const LoadingOrb = ({ size = 46 }) => {
+    const orbSize = 10;
+    const orbitRadius = size * 0.32;
+
+    return (
+        <Box
+            position="relative"
+            display="flex"
+            alignItems="center"
+            justifyContent="center"
+            w={`${size}px`}
+            h={`${size}px`}
+            borderRadius="full"
+            overflow="hidden"
+            animation={`${fadeIn} 0.3s ease-out forwards`}
+        >
+            {/* Central glowing core */}
+            <Box
+                position="absolute"
+                borderRadius="full"
+                w={`${size * 0.35}px`}
+                h={`${size * 0.35}px`}
+                bg="radial-gradient(circle, #FF8A50 0%, #FF6B35 50%, transparent 70%)"
+                animation={`${corePulse} 1.5s ease-in-out infinite`}
+                boxShadow="0 0 15px rgba(255, 107, 53, 0.6)"
+            />
+
+            {/* Orbit ring 1 - outer */}
+            <Box
+                position="absolute"
+                w={`${orbitRadius * 2}px`}
+                h={`${orbitRadius * 2}px`}
+                borderRadius="full"
+                animation={`${swirlOrbit} 2s linear infinite`}
+            >
+                <Box
+                    position="absolute"
+                    top="0"
+                    left="50%"
+                    transform="translateX(-50%)"
+                    w={`${orbSize}px`}
+                    h={`${orbSize}px`}
+                    borderRadius="full"
+                    bg="linear-gradient(135deg, #FF6B35, #FF8A50)"
+                    boxShadow="0 0 8px rgba(255, 107, 53, 0.8)"
+                    animation={`${orbPulse} 1s ease-in-out infinite`}
+                />
+            </Box>
+
+            {/* Orbit ring 2 - middle */}
+            <Box
+                position="absolute"
+                w={`${orbitRadius * 1.5}px`}
+                h={`${orbitRadius * 1.5}px`}
+                borderRadius="full"
+                animation={`${swirlOrbitReverse} 1.5s linear infinite`}
+            >
+                <Box
+                    position="absolute"
+                    bottom="0"
+                    left="50%"
+                    transform="translateX(-50%)"
+                    w={`${orbSize * 0.8}px`}
+                    h={`${orbSize * 0.8}px`}
+                    borderRadius="full"
+                    bg="linear-gradient(135deg, #FFB74D, #FF8A50)"
+                    boxShadow="0 0 6px rgba(255, 183, 77, 0.8)"
+                    animation={`${orbPulse} 1.2s ease-in-out infinite 0.3s`}
+                />
+            </Box>
+
+            {/* Orbit ring 3 - inner */}
+            <Box
+                position="absolute"
+                w={`${orbitRadius * 1}px`}
+                h={`${orbitRadius * 1}px`}
+                borderRadius="full"
+                animation={`${swirlOrbit} 1s linear infinite`}
+            >
+                <Box
+                    position="absolute"
+                    top="50%"
+                    right="0"
+                    transform="translateY(-50%)"
+                    w={`${orbSize * 0.6}px`}
+                    h={`${orbSize * 0.6}px`}
+                    borderRadius="full"
+                    bg="linear-gradient(135deg, #FFCC80, #FFB74D)"
+                    boxShadow="0 0 5px rgba(255, 204, 128, 0.8)"
+                    animation={`${orbPulse} 0.8s ease-in-out infinite 0.5s`}
+                />
+            </Box>
+        </Box>
+    );
+};
